requirements.txt
cassandra-driver==2.1.1
